added sorting/filtering/pagination to test fetching on ui

+++ generation-service/src/main/java/com/example/generation_service/dto/tests/TestsResponseDto.java
@@ -12,7 +12,7 @@ public class TestsResponseDto {
 
   private final List<TestResponseDto> tests;
   private final int totalPages;
-  private final int totalElements;
+  private final long totalElements;
 
   @Getter
   @Builder
